Modified the script to work with MVC structure for better separation and cleaner project
