+++ controllers/bitcoinController.js
@@ -1,0 +1,11 @@
+const bitcoinService = require('../services/bitcoinService');
+
+async function trackBitcoin() {
+  try {
+    await bitcoinService.fetchAndStore();
+  } catch (err) {
+    console.error("Controller error:", err.message);
+  }
+}
+
+module.exports = { trackBitcoin };
